rename pg_bouncer to pgbouncer
