+++ compute_tools/src/lib.rs
@@ -21,7 +21,7 @@ pub mod metrics;
 mod migration;
 pub mod monitor;
 pub mod params;
-pub mod pg_bouncer;
+pub mod pgbouncer;
 pub mod pg_helpers;
 pub mod rsyslog;
 pub mod spec;
